fix: review api error

+++ backendServices/review.ts
@@ -18,7 +18,7 @@ export const postReview = (review: Review): Promise<{done: boolean}> => {
   $ip_address: String!
   $session_id: String!
   $station_id: Int!
-  $stars: Float!
+  $stars: Int!
   $message: String
 ) {
   insert_reviews_one(
